update dashboard info
