+++ frontend/src/App.js
@@ -8,7 +8,7 @@ import Onboarding from './features/onboarding/Onboarding';
 import SignUp from './features/signup/SignUp';
 import Login from './features/login/Login'
 import Token from './features/token/Token'
-import Dashboard from './layouts/dashboard'
+import Dashboard from './layouts/personalinfo'
 
 // @mui material components
 import { ThemeProvider } from "@mui/material/styles";
