+++ frontend/src/layouts/personalinfo/components/Emergency/index.js
@@ -6,37 +6,29 @@ import MDBox from "components/MDBox";
 
 // Material Dashboard 2 React components
 import ProfileInfoCard from "examples/Cards/InfoCards/ProfileInfoCard";
-import ReportsLineChart from "examples/Charts/LineCharts/ReportsLineChart";
-import ReportsBarChart from "examples/Charts/BarCharts/ReportsBarChart";
-
-// Data
-import reportsLineChartData from "layouts/dashboard/data/reportsLineChartData";
 
 // @mui icons
-import PageviewIcon from '@mui/icons-material/Pageview';
+import ForumIcon from '@mui/icons-material/Forum';
 
 
-function DocumentSection() {
-
-    const { sales, tasks } = reportsLineChartData;
+function Emergency() {
 
     return (
-
         <MDBox>
             <Grid container spacing={3}>
                 <Grid item xs={12} md={12} lg={12}>
                     <MDBox mb={1.5}>
                         <ProfileInfoCard
-                            title="Documents"
+                            title="Emergency Contact"
                             info={{
-                            upload1: "Symbol | File Name1 | 1/5/2017",
-                            upload2: "Symbol | File Name2 | 12/28/2012",
-                            upload3: "Symbol | File Name3 | 1/21/2009",
+                            fullName: "John Doe",
+                            phone: "(111) 111 - 1111",
+                            address: "1234 Street Name Philadelphia, PA 07012",
                             }}
                             social={[
                                 {
-                                    link: "https://www.google.com/docs/about/",
-                                    icon: <PageviewIcon />,
+                                    link: "https://www.wechat.com/",
+                                    icon: <ForumIcon />,
                                     color: "github",
                                 },
                             ]}
@@ -50,4 +42,4 @@ function DocumentSection() {
     )
 }
 
-export default DocumentSection;
+export default Emergency;
