+++ frontend/src/layouts/personalinfo/components/Profile/index.js
@@ -21,7 +21,9 @@ import ProfileInfoCard from "examples/Cards/InfoCards/ProfileInfoCard";
 import Grid from "@mui/material/Grid";
 
 // @mui icons
-import LocationOnIcon from '@mui/icons-material/LocationOn';
+import FacebookIcon from '@mui/icons-material/Facebook';
+import TwitterIcon from '@mui/icons-material/Twitter';
+import InstagramIcon from '@mui/icons-material/Instagram';
 
 function Address() {
 
@@ -29,27 +31,38 @@ function Address() {
 
     <MDBox>
       <Grid container spacing={3}>
-        <Grid item xs={12} lg={6}>
+        <Grid item xs={12} lg={12}>
           <ProfileInfoCard
             title="User Profile"
             info={{
-              addressLine1: "1111 Street Name",
-              addressLine2: "Apt #123",
-              city: "Philadelphia",
-              state: "PA",
-              zip: "07123",
+              fullName: "Natasha Lee",
+              preferredName: "Natasha",
+              birthDate: "05 June 1996",
+              age: "26",
+              gender: "Female",
+              SSN: "***-***-1234",
             }}
             social={[
               {
-                link: "https://www.google.com/maps",
-                icon: <LocationOnIcon />,
-                color: "github",
+                link: "https://www.facebook.com",
+                icon: <FacebookIcon />,
+                color: "facebook",
+              },
+              {
+                link: "https://www.twitter.com",
+                icon: <TwitterIcon />,
+                color: "twitter",
+              },
+              {
+                link: "https://www.instagram.com",
+                icon: <InstagramIcon />,
+                color: "instagram",
               },
             ]}
             action={{ route: "", tooltip: "Edit Profile" }}
           />
         </Grid>
-        <Grid item xs={12} lg={6}>
+        {/* <Grid item xs={12} lg={6}>
           <ProfileInfoCard
             title="Secondary Address"
             info={{
@@ -68,7 +81,7 @@ function Address() {
             ]}
             action={{ route: "", tooltip: "Edit Profile" }}
           />
-        </Grid>
+        </Grid> */}
       </Grid>
     </MDBox>
   );
